Created util package to separate utilities such as validity checking, renamed thinkdeep type to model
