+++ packages/data-collector/package.json
@@ -17,7 +17,8 @@
   },
   "dependencies": {
     "@thinkdeep/attach-exit-handler": "^1.0.0",
-    "@thinkdeep/type": "file:./../type",
+    "@thinkdeep/model": "file:./../model",
+    "@thinkdeep/util": "file:./../util",
     "commander": "^9.0.0",
     "kafkajs": "^2.1.0",
     "log4js": "^6.3.0",
