+++ packages/deep-microservice-analysis/package.json
@@ -30,7 +30,8 @@
     "@thinkdeep/apollo-datasource-neo4j": "^0.0.1-alpha.7",
     "@thinkdeep/attach-exit-handler": "^1.0.0",
     "@thinkdeep/get-public-ip": "^1.0.0",
-    "@thinkdeep/type": "file:../type",
+    "@thinkdeep/model": "file:../model",
+    "@thinkdeep/util": "file:./../util",
     "apollo-datasource-mongodb": "^0.5.2",
     "apollo-server": "3.8.0",
     "apollo-server-express": "^3.5.0",
