+++ packages/deep-microservice-configuration/package.json
@@ -29,7 +29,7 @@
     "@apollo/subgraph": "^0.1.5",
     "@thinkdeep/attach-exit-handler": "^1.0.0",
     "@thinkdeep/get-public-ip": "^1.0.0",
-    "@thinkdeep/type": "file:../type",
+    "@thinkdeep/model": "file:../model",
     "apollo-datasource-mongodb": "^0.5.2",
     "apollo-server": "3.8.0",
     "apollo-server-express": "^3.5.0",
