+++ packages/deep-microservice-subscription/Dockerfile
@@ -10,7 +10,7 @@ COPY src/ ./src
 
 COPY package.json ./
 
-COPY ./node_modules/@thinkdeep/type ../type/
+COPY ./node_modules/@thinkdeep/model ../model/
 
 RUN yarn install
