+++ packages/deep-microservice-subscription/src/schema.js
@@ -1,6 +1,6 @@
 import {gql} from 'apollo-server';
 
-import {EconomicEntityFactory, EconomicEntityType} from '@thinkdeep/type';
+import {EconomicEntityFactory, EconomicEntityType} from '@thinkdeep/model';
 
 const economicEntity = EconomicEntityFactory.economicEntity({
   name: 'dummy',
