+++ packages/model/README.md
@@ -1,0 +1,11 @@
+# `@thinkdeep/model`
+
+> TODO: description
+
+## Usage
+
+```
+const type = require('@thinkdeep/model');
+
+// TODO: DEMONSTRATE API
+```
