+++ packages/model/src/economic-entity-type.js
