+++ packages/model/src/index.js
@@ -4,5 +4,3 @@ export {
 } from './economic-entity.js';
 export {EconomicEntityFactory} from './economic-entity-factory.js';
 export {EconomicEntityType} from './economic-entity-type.js';
-export {validString} from './string.js';
-export {validDate} from './date.js';
