+++ packages/util/package.json
@@ -1,0 +1,37 @@
+{
+  "name": "@thinkdeep/util",
+  "version": "0.0.9-alpha.1",
+  "description": "Utility functions.",
+  "keywords": [
+    "NodeJS",
+    "MomentJS"
+  ],
+  "author": "Hayden McParlane <hayden.mcparlane@gmail.com>",
+  "homepage": "https://github.com/ThinkDeepTech/thinkdeep#readme",
+  "license": "GPL-3.0",
+  "main": "src/index.js",
+  "type": "module",
+  "dependencies": {
+    "moment": "^2.29.4"
+  },
+  "directories": {
+    "lib": "src",
+    "test": "test"
+  },
+  "files": [
+    "src"
+  ],
+  "publishConfig": {
+    "access": "public"
+  },
+  "repository": {
+    "type": "git",
+    "url": "git+https://github.com/ThinkDeepTech/thinkdeep.git"
+  },
+  "scripts": {
+    "test": "echo \"Error: run tests from root\" && exit 1"
+  },
+  "bugs": {
+    "url": "https://github.com/ThinkDeepTech/thinkdeep/issues"
+  }
+}
